Avoid including querystring in file extension
