+++ droid-command-line/src/test/java/uk/gov/nationalarchives/droid/command/archive/WarcArchiveContentIdentifierTest.java
@@ -59,14 +59,11 @@ public class WarcArchiveContentIdentifierTest {
     private WarcArchiveContentIdentifier warcArchiveContentIdentifier;
     private ContainerSignatureDefinitions containerSignatureDefinitions;
     private String standardSignatures =
-            //"src/test/resources/signatures/DROID_SignatureFile_V78.xml";
-    "/home/dev/github/droid_seamang/droid-command-line/src/test/resources/signatures/DROID_SignatureFile_V78.xml";
+            "src/test/resources/signatures/DROID_SignatureFile_V78.xml";
     private String containerSignatures =
-            //"src/test/resources/signatures/container-signature-20140922.xml";
-        "/home/dev/github/droid_seamang/droid-command-line/src/test/resources/signatures/container-signature-20140922.xml";
+            "src/test/resources/signatures/container-signature-20140922.xml";
     private String warcFile =
-            //"src/test/resources/testfiles/expanded.warc";
-    "/home/dev/github/droid_seamang/droid-command-line/src/test/resources/testfiles/expanded.warc";
+            "src/test/resources/testfiles/expanded.warc";
     
     @Before
     public void setUp() throws CommandExecutionException {
