+++ droid-core-interfaces/src/main/java/uk/gov/nationalarchives/droid/core/interfaces/archive/WarcArchiveHandler.java
@@ -35,13 +35,10 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
 import java.net.URL;
-import java.util.List;
-import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.HashMap;
 import java.util.Date;
-import java.util.regex.Pattern;
 
 import org.jwat.warc.WarcHeader;
 import org.jwat.warc.WarcReader;
@@ -50,12 +47,8 @@ import org.jwat.warc.WarcRecord;
 
 import uk.gov.nationalarchives.droid.core.interfaces.ResourceId;
 import uk.gov.nationalarchives.droid.core.interfaces.RequestIdentifier;
-import uk.gov.nationalarchives.droid.core.interfaces.IdentificationResultImpl;
-import uk.gov.nationalarchives.droid.core.interfaces.AsynchDroid;
 import uk.gov.nationalarchives.droid.core.interfaces.IdentificationRequest;
-import uk.gov.nationalarchives.droid.core.interfaces.ResultHandler;
 import uk.gov.nationalarchives.droid.core.interfaces.resource.RequestMetaData;
-
 
 
 /**
@@ -63,15 +56,11 @@ import uk.gov.nationalarchives.droid.core.interfaces.resource.RequestMetaData;
  * @author gseaman
  *
  */
-public class WarcArchiveHandler implements ArchiveHandler {
+public class WarcArchiveHandler extends WebArchiveHandler implements ArchiveHandler {
     /**
-     * Save importing all the http codes
+     * Used to generate URIs within Warc file
      */
-    protected static final int HTTP_ACCEPTED = 200;
-    private AsynchDroid droidCore;
-    private IdentificationRequestFactory factory;
-    private ResultHandler resultHandler;
-
+    protected static final String WEB_ARCHIVE_TYPE = "warc";
 
     /**
      * {@inheritDoc}
@@ -98,29 +87,6 @@ public class WarcArchiveHandler implements ArchiveHandler {
         }
     }
     
-
-    
-
-    /**
-     * @param factory the factory to set
-     */
-    public final void setFactory(IdentificationRequestFactory factory) {
-        this.factory = factory;
-    }
-
-    /**
-     * @param droidCore the droidCore to set
-     */
-    public final void setDroidCore(AsynchDroid droidCore) {
-        this.droidCore = droidCore;
-    }
-
-    /**
-     * @param resultHandler the resultHandler to set
-     */
-    public final void setResultHandler(ResultHandler resultHandler) {
-        this.resultHandler = resultHandler;
-    }
 
     /**
      * Adapts the JWAT ByteCountingPushbackInputStream to generate warc entries.
@@ -189,35 +155,8 @@ public class WarcArchiveHandler implements ArchiveHandler {
                 time == null ? null : time.getTime(),
                 entryName);
 
-        RequestIdentifier identifier =
-                new RequestIdentifier(ArchiveFileUtils.toWarcUri(parentName, entryName));
-        identifier.setAncestorId(originatorNodeId);
-        identifier.setParentResourceId(correlationId);
-        IdentificationRequest request = factory.newRequest(metaData, identifier);
-        request.open(entry.getPayloadContent()); // get the inputstream back from the record, and write it out to file
-        droidCore.submit(request);
-    }
-
-    /**
-     *
-     * @param parentName
-     * @param entry
-     * @param entryName
-     * @param correlationId
-     * @return
-     */
-    private ResourceId submitDirectory(final URI parentName,
-                                       WarcRecord entry, String entryName, ResourceId correlationId) {
-        IdentificationResultImpl result = new IdentificationResultImpl();
-
-        RequestMetaData metaData = new RequestMetaData(null, null, entryName);
-
-        RequestIdentifier identifier = new RequestIdentifier(
-            ArchiveFileUtils.toWarcUri(parentName, entry.getHttpHeader().requestUri));
-
-        result.setRequestMetaData(metaData);
-        result.setIdentifier(identifier);
-        return resultHandler.handleDirectory(result, correlationId, false);
+        super.submit(WEB_ARCHIVE_TYPE, metaData, parentName,
+                entry.getPayloadContent(), correlationId, originatorNodeId);
     }
 
 
@@ -250,88 +189,20 @@ public class WarcArchiveHandler implements ArchiveHandler {
             String entryName = entryPath.substring(entryPath.lastIndexOf('/') + 1);
             String prefixPath = entryUri.substring(0, entryUri.length() - entryName.length());
             ResourceId correlationId = parentId; // by default, files are correlated to the parent.
+            String requestUri = entry.getHttpHeader().requestUri;
 
             // If there is a path, get the actual correlation id for its parent folder:
             if (!prefixPath.isEmpty()) {
                 correlationId = directories.get(prefixPath);
                 // If we haven't seen the path before, add the ancestor folders not yet seen:
                 if (correlationId == null) {
-                    correlationId = processAncestorFolders(prefixPath, entry);
+                    correlationId = processAncestorFolders(WEB_ARCHIVE_TYPE, prefixPath,
+                            requestUri, parentId, parentName, directories);
                 }
             }
 
             submit(entry, entryName, parentName, in, correlationId, originatorNodeId);
         }
 
-
-
-
-        /**
-         * Finds the longest path which has been seen before (if any),
-         * and adds all the subsequent folders which haven't been seen.
-         * @param prefixPath the path of
-         */
-        private ResourceId processAncestorFolders(String url, WarcRecord entry) {
-            // Split the path string into a list of ancestor paths:
-            List<String> urls = getAncestorUrls(url);
-            ResourceId longestParentId = parentId;
-
-            // Find the longest path we *have* seen before (if any):
-            // (ancestor paths are ordered longest first)
-            int longestSeenBefore = urls.size();
-            ResourceId correlationId = null;
-            for (int urlIndex = 0; urlIndex < urls.size(); urlIndex++) {
-                correlationId = directories.get(urls.get(urlIndex));
-                if (correlationId != null) {
-                    longestSeenBefore = urlIndex;
-                    longestParentId = correlationId;
-                    break;
-                }
-            }
-
-            // Add any that haven't yet been processed (from shortest to longest)
-            Pattern p = Pattern.compile("/");
-            for (int urlIndex = longestSeenBefore - 1; urlIndex >= 0; urlIndex--) {
-                final String urlName = urls.get(urlIndex);
-                String dirName = urlName;
-                String[] dirs = p.split(urlName);
-                // extract the last directory name if any
-                if (!"".equals(dirs[dirs.length - 1])) {
-                    dirName = dirs[dirs.length - 1];
-                }
-                longestParentId = submitDirectory(parentName, entry, dirName, longestParentId);
-                directories.put(urlName, longestParentId);
-            }
-
-            return longestParentId;
-        }
-
     }
-    /**
-     *
-     * @param url Url as string up to last containing directory
-     * @return String[] a string array containing the parent folders, each of
-     *         which is a url string in its own right (not just the names of each individual folder)
-     */
-    public static List<String> getAncestorUrls(String url) {
-        String schemeSeparator = "://";
-        ArrayList<String> urls = new ArrayList<String>();
-        if (url != null && !url.isEmpty()) {
-            String urlPath = url.substring(url.indexOf(schemeSeparator) + schemeSeparator.length());
-            String urlScheme = url.substring(0, url.indexOf(schemeSeparator) + schemeSeparator.length());
-            char separator = '/'; // even on Windows
-            // allow for urls without trailing slash
-            int lastSeparator = urlPath.length();
-            if (separator == urlPath.charAt(urlPath.length() - 1)) {
-                lastSeparator -= 1;
-            }
-            while (lastSeparator >= 0) {
-                urlPath = urlPath.substring(0, lastSeparator);
-                urls.add(urlScheme + urlPath + String.valueOf(separator));
-                lastSeparator = urlPath.lastIndexOf(separator);
-            }
-        }
-        return urls;
-    }
-
 }
